Fix crash in AddActionDialog.
Fix a crash that occurred when selecting an action equal to the action
that opened the AddActionDialog. The problem was because both actions
were trying to use the same EditorWidget at the same time. The solution
was to create a new EditorWidget in the AddActionDialog when requested
instead of reusing the one already created.