Change variable syntax within text.
Now to specify a variable it's also required to use curly braces ({})
for delimiting the variable name. For example, what before would be
written as "Hello $name", now is written "Hello {$name}".

This allows more freedom in writting text, because it makes more
difficult to misinterpret a variable and also allows to put the variable
together with other text more easily.